refactor(newSpecial):기본 레이아웃 구조 수정(스티키 메뉴 관련)
