+++ src/AppProvider.tsx
@@ -1,4 +1,5 @@
 import { Suspense, useState } from 'react'
+import { BrowserRouter } from 'react-router-dom'
 import { ErrorBoundary } from 'react-error-boundary'
 import {
   QueryCache,
@@ -25,10 +26,12 @@ export const AppProvider = ({ children }: { children: React.ReactNode }) => {
   )
 
   return (
-    <QueryClientProvider client={queryClient}>
-      <ErrorBoundary fallback={<div>이용에 불편을 드려 죄송합니다.</div>}>
-        <Suspense fallback={<div>로딩중...</div>}>{children}</Suspense>
-      </ErrorBoundary>
-    </QueryClientProvider>
+    <BrowserRouter>
+      <QueryClientProvider client={queryClient}>
+        <ErrorBoundary fallback={<div>이용에 불편을 드려 죄송합니다.</div>}>
+          <Suspense fallback={<div>로딩중...</div>}>{children}</Suspense>
+        </ErrorBoundary>
+      </QueryClientProvider>
+    </BrowserRouter>
   )
 }
