+++ src/assets/scss/base/_typography.scss
@@ -1,4 +1,4 @@
-@import '../../assets/font/pretendard/pretendard.css';
+@import '../../font/pretendard/pretendard.css';
 
 @font-face {
   font-family: 'NeoDunggeunmo';
