+++ src/assets/scss/pages/_main.scss
@@ -80,7 +80,7 @@ $star-sizes: (
 //------------------------------------------//
 // 메인
 //------------------------------------------//
-.main {
+#main {
   background: get-color(neutral, EerieBlack);
   //---------------- 키비주얼 영역 ----------------//
   .sc_hero {
