+++ src/domains/_common/components/layouts/PageContainer.tsx
@@ -1,0 +1,3 @@
+export const PageContainer = ({ children }: { children: React.ReactNode }) => {
+  return <main className="page_container">{children}</main>
+}
